Release 0.7.3.

+++ processing/src/generate_tiles_for_image.sh
@@ -81,14 +81,14 @@ function generate_tiles() {
 # tile is smaller or equal to 64x64.
 function scale_and_generate_tiles() {
     tiles_dir=$1$2/$3;
-    if [ ! -d "tiles_dir" ]
+    if [[ ! -d "${tiles_dir}" ]]
     then
         #echo "    Directory $tiles_dir for $2x$2 tiles at $3% scale does not exists...";
         mkdir -p ${tiles_dir};
     fi;
     
     # Uses ImageMagick to scale the image
-    if [ ! -z "$3" ]
+    if [[ ! -z "$3" ]]
     then
         (cd ${tiles_dir}; convert ../../original.${extension} -resize $3% scaled.${extension}; )
     fi;
@@ -105,7 +105,7 @@ function scale_and_generate_tiles() {
 function generate_tiles_set() {
     checksum_dir=$1;
 
-    if [ "$extension" = "dcm" ]
+    if [[ "${extension}" = "dcm" ]]
     then
         #echo "    Converting DICOM to preferred image format...";
         (cd ${checksum_dir}; convert -define dcm:display-range=reset original.${extension} -normalize original.${pref_format}; )
@@ -118,13 +118,16 @@ function generate_tiles_set() {
     conversion_status=$?
 
     # Delete original media file since it is archived in the source directory.
-    (cd ${checksum_dir}; rm -f original.${extension}; )
+    if [[ "${extension}" != "${pref_format}" ]]
+    then
+        (cd ${checksum_dir}; rm -f original.${extension}; )
+    fi;
 
     # Check if converion to JPEG was successful; otherwise, stop the tiling.
-    if [[ ${conversion_status} != 0 ]];
+    if [[ ${conversion_status} -ne 0 ]]
     then
         echo "    Failed to convert media file to JPEG... will abort";
-        exit ${conversion_status};
+        exit 1;
     fi;
 
     # we convert all images to preferred image format before processing
@@ -135,12 +138,15 @@ function generate_tiles_set() {
     #
     # NOTE: we are assuming TIF files may contain the same image at
     # different zooms; but not different images.
-    if [ `find ${checksum_dir} -iname "*.${pref_format}" -type f | wc -l` -gt 1 ]
+    if [[ `find ${checksum_dir} -maxdepth 1 -iname "*.${pref_format}" -type f | wc -l` -gt 1 ]]
     then
-        largest_file=`find ${checksum_dir} -iname "*.${pref_format}" -type f -print0 | xargs -0 du -b | sort -nr | head -n 1 | cut -f 2`;
-        mv ${largest_file} ${checksum_dir}original.${pref_format};
-        
-        # Delete all of the other extracted images
+        largest_file=`find ${checksum_dir} -maxdepth 1 -iname "*.${pref_format}" -type f -print0 | xargs -0 du -b | sort -nr | head -n 1 | cut -f 2`;
+        if [[ "$largest_file" != "${checksum_dir}original.${pref_format}" ]]
+        then
+            mv -f ${largest_file} ${checksum_dir}original.${pref_format};
+        fi;
+
+        # Delete all of the files except for the original JPEG image to tile
         find ${checksum_dir} ! -iname "original.${pref_format}" -type f -delete;
     fi;
 
@@ -163,24 +169,23 @@ function generate_tiles_set() {
     (cd ${checksum_dir}; rm -f original.${pref_format}; )
 }
 
-if [ ! $# -eq 4 ]
+if [[ ! $# -eq 4 ]]
 then
     echo "Usage:";
     echo "    generate_tiles_for_image.sh <file> <destination> <sizes> <scales>";
     echo "";
     echo " E.g., generate_tiles_for_image.sh example.jpg tiles 256 10,25,50,75,100";
     echo "";
-    echo " file - Original image file to process.";
+    echo "        file - Original image file to process.";
     echo " destination - Destination directory to store image tiles set.";
-    echo " sizes - Comma separated list of tile sizes (in pixels).";
-    echo " scales - Comma separated list of scales (in percentage).";
+    echo "       sizes - Comma separated list of tile sizes (in pixels).";
+    echo "      scales - Comma separated list of scales (in percentage).";
     exit 1;
 fi;
 
-if [ ! -f ${file_to_process} ]
+if [[ ! -f ${file_to_process} ]]
 then
-    echo "ERROR: Supplied image file does not exists...";
-    echo "Will abort image tile generation...";
+    echo "Abort tiling... ${file_to_process} does not exists...";
     exit 1;
 fi;
 
@@ -192,22 +197,15 @@ case ${tiles_directory} in
         ;;
 esac
 
-if [ ! -d ${tiles_directory} ]
+if [[ ! -d ${tiles_directory} ]]
 then
-    echo "Supplied target directory does not exists...";
-    echo "    Creating directory...";
     mkdir -p ${tiles_directory};
 fi;
 
 basename=$(basename "${file_to_process}");
 extension="${file_to_process##*.}";
 filename=`basename ${basename} .${extension}`;
-#echo "    Basename: ${basename}";
-#echo "    Filename: ${filename}";
-#echo "    Extension: ${extension}";
-
 checksum=`${sha_cmd} ${file_to_process} | cut -d " " -f 1`;
-#echo "    Checksum: ${checksum}";
 
 # We could just use the checksum, but the number of directories
 # in this flat representation could reach the filesystem limit.
@@ -225,22 +223,16 @@ checksum=`${sha_cmd} ${file_to_process} | cut -d " " -f 1`;
 # b5ea/f562/7bea/e587/bc77/9b68/ec8f/41b3/0caa/4015
 #
 checksum_hierarchy=`echo ${checksum} | sed 's/.\{4\}/&\//g'`
-#echo ${checksum_hierarchy};
 checksum_dir=${tiles_directory}${checksum_hierarchy};
 checksum_filepath=${checksum_dir}original.${extension};
 
-# Check if the file to process exists.
-if [ -e "${file_to_process}" ]
+if [[ -d "${checksum_dir}" ]]
 then
-    # Check if we have already generated the tiles
-    if [ -d "${checksum_dir}" ]
-    then
-        echo "        Tile directory ${checksum_dir} already exists... will skip tiling";
-    else
-        mkdir -p ${checksum_dir};
-        cp -f ${file_to_process} ${checksum_filepath};
-        generate_tiles_set ${checksum_dir} ${tile_sizes} ${image_scales};
-    fi;
-else
-    echo "        File ${file_to_process} does not exists... will abort tiling";
+    echo "    ${checksum_dir} exists... will delete and redo tiling";
+    rm -Rf ${checksum_dir};
 fi;
+
+echo "Processing ${file_to_process}...";
+mkdir -p ${checksum_dir};
+cp -f ${file_to_process} ${checksum_filepath};
+generate_tiles_set ${checksum_dir} ${tile_sizes} ${image_scales};
